requirements.txt
appdirs==1.4.4
asgiref==3.2.7
attrs==19.3.0
autopep8==1.5.4
boto3==1.14.20
botocore==1.17.20
certifi==2020.4.5.1
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
Deprecated==1.2.10
distlib==0.3.1
Django==3.0.5
django-debug-toolbar==3.2.1
django-filter==2.4.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.7.1
docutils==0.15.2
Faker==8.10.3
filelock==3.0.12
flake8==3.8.4
Flask==1.1.2
gitdb==4.0.5
GitPython==3.1.3
gunicorn==20.0.4
identify==1.4.24
idna==2.9
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
mccabe==0.6.1
mixer==7.1.2
more-itertools==8.2.0
nodeenv==1.4.0
numpy==1.19.5
packaging==20.3
pdfminer.six==20200104
Pillow==7.0.0
pluggy==0.13.1
pre-commit==2.6.0
psycopg2-binary==2.8.5
py==1.8.1
pycodestyle==2.6.0
pycryptodome==3.9.7
pyflakes==2.2.0
PyGithub==1.51
PyJWT==2.1.0
pyparsing==2.4.7
pytest==5.4.1
pytest-django==3.9.0
pytest-timeout==1.3.4
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
six==1.14.0
smmap==3.0.4
sorl-thumbnail==12.6.3
sortedcontainers==2.1.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.1
urllib3==1.25.9
virtualenv==20.0.27
wcwidth==0.1.9
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
